pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "threadmesh"
version = "0.1.0"
description = "High-quality FEA/CFD mesh optimizer with full GUI"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.11"

authors = [
    { name = "ThreadMesh Contributors" }
]

keywords = ["mesh", "FEA", "CFD", "optimization", "finite element", "meshing"]

classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Manufacturing",
    "Operating System :: OS Independent",
]

dependencies = [
    "PySide6>=6.6.0",
    "vtk>=9.3.0",
    "meshio>=5.3.0",
    "h5py>=3.10.0",
    "gmsh>=4.12.0",
    "pymeshlab>=2023.12",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "numba>=0.59.0",
    "matplotlib>=3.8.0",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "nvidia-ml-py",
]

[project.optional-dependencies]
cuda = ["cupy-cuda12x"]
opencl = ["pyopencl"]
gpu = ["cupy-cuda12x", "pyopencl"]

[project.scripts]
threadmesh = "threadmesh.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["threadmesh*"]
